doubles tested

+++ bubble-sort/bubblesort.h
@@ -1,3 +1,3 @@
-typedef float (*compare)(void* first, void* second) ;
+typedef int (*compare)(void* first, void* second) ;
 
 void bubbleSort(void* base, int numberOfElements, int elementSize , compare comp);
